+++ DICOM/IO/Buffer/FileByteBuffer.cs
@@ -6,35 +6,26 @@ namespace Dicom.IO.Buffer
     public sealed class FileByteBuffer : IByteBuffer
     {
 
-        public FileByteBuffer(IFileReference file, long position, uint length)
+        public FileByteBuffer(IFileReference file, long position, long length)
         {
             File = file;
             Position = position;
             Size = length;
         }
 
-        public bool IsMemory
-        {
-            get => false;
-        }
+        public bool IsMemory => false;
 
         public IFileReference File { get; private set; }
 
         public long Position { get; private set; }
 
-        public uint Size { get; private set; }
+        public long Size { get; private set; }
 
-        public byte[] Data
-        {
-            get
-            {
-                return File.GetByteRange((int)Position, (int)Size);
-            }
-        }
+        public byte[] Data => File.GetByteRange(Position, (int)Size);
 
-        public byte[] GetByteRange(int offset, int count)
+        public byte[] GetByteRange(long offset, int count)
         {
-            return File.GetByteRange((int)Position + offset, count);
+            return File.GetByteRange(Position + offset, count);
         }
 
     }